Refactor model task paths and add new evaluation scripts
- Updated model task paths to change "cnn_v2" to "cnn_v2_pair" for clarity.
- Introduced a new shell script `plot_eval_isolated_mmus.sh` for plotting evaluation scores for the Mmus species.
- Added unit tests for general model SNPR functionalities, including species curve computations and interactive HTML generation.
- Enhanced hyperparameter search tests to validate the new "cnn_v2_pair" model.
- Modified mixed pair negative data processing tests to ensure correct directory structure and functionality.
- Updated model task path tests to include the new "cnn_v2_pair" model.
- Added tests for new functions in the raw data notebook library related to test intron and site label counts.
- Implemented additional tests for SNPR curves, including ROC curve generation and agreement metrics.
- Created tests for transcript truth simple models, ensuring proper feature extraction and evaluation.

+++ src/models/bert_drosophila.py
@@ -598,7 +598,7 @@ def score_sequences(
         batch_ids = input_ids[i : i + batch_size].to(device)
         batch_mask = attn_mask[i : i + batch_size].to(device)
         logits = model(batch_ids, batch_mask)
-        probs = torch.sigmoid(logits).cpu().numpy()
+        probs = torch.sigmoid(logits).float().cpu().numpy()
         all_probs.append(probs)
 
     return np.concatenate(all_probs) if all_probs else np.array([])
