+++ src/models/cnn_common.py
@@ -428,7 +428,7 @@ def score_sequences(
             amp_context = nullcontext()
         with amp_context:
             logits = model(batch_x)
-        probs = torch.sigmoid(logits).detach().cpu().numpy()
+        probs = torch.sigmoid(logits).float().detach().cpu().numpy()
         all_probs.append(probs)
 
     return np.concatenate(all_probs)
